moved the proximity indicator into its own widget
